delete projects for older Play versions, bring support for Play 2.9 and 3.0
