+++ project/plugins.sbt
@@ -1,3 +1,2 @@
 addSbtPlugin("com.geirsson" % "sbt-ci-release" % "1.5.7")
-addSbtPlugin("com.lightbend.sbt" % "sbt-javaagent" % "0.1.4")
 addSbtPlugin("de.heikoseeberger" % "sbt-header" % "5.6.0")
